+++ iroh-relay/src/protos.rs
@@ -1,7 +1,7 @@
 //! Protocols used by the iroh-relay
 
+pub mod common;
 pub mod disco;
 pub mod handshake;
 pub mod relay;
-pub mod send_recv;
 pub mod streams;
